+++ VocalRaidAssistant.lua
@@ -6,6 +6,7 @@ VRA = LibStub("AceAddon-3.0"):NewAddon(addon, addonName, "AceConsole-3.0", "AceE
 VRA.L = LibStub("AceLocale-3.0"):GetLocale(addonName)
 VRA.AC = LibStub("AceConfig-3.0")
 VRA.ACD = LibStub("AceConfigDialog-3.0")
+VRA.AG = LibStub("AceGUI-3.0")
 VRA.ACR = LibStub("AceConfigRegistry-3.0")
 VRA.ACDBO = LibStub("AceDBOptions-3.0")
 VRA.EXP = LibStub("AceSerializer-3.0")